ast_results/open-keychain_open-keychain/OpenKeychain/src/main/java/org/sufficientlysecure/keychain/operations/BackupOperation.java
// isComment
package org.sufficientlysecure.keychain.operations;

import java.io.BufferedOutputStream;
import java.io.FileNotFoundException;
import java.io.IOException;
import java.io.InputStream;
import java.io.OutputStream;
import java.text.SimpleDateFormat;
import java.util.Date;
import java.util.List;
import java.util.Locale;
import java.util.concurrent.atomic.AtomicBoolean;
import java.util.regex.Pattern;
import android.content.Context;
import android.net.Uri;
import android.support.annotation.NonNull;
import android.support.annotation.Nullable;
import org.bouncycastle.bcpg.ArmoredOutputStream;
import org.sufficientlysecure.keychain.Constants;
import org.sufficientlysecure.keychain.R;
import org.sufficientlysecure.keychain.daos.KeyRepository;
import org.sufficientlysecure.keychain.daos.KeyRepository.NotFoundException;
import org.sufficientlysecure.keychain.model.SubKey.UnifiedKeyInfo;
import org.sufficientlysecure.keychain.operations.results.ExportResult;
import org.sufficientlysecure.keychain.operations.results.OperationResult.LogType;
import org.sufficientlysecure.keychain.operations.results.OperationResult.OperationLog;
import org.sufficientlysecure.keychain.operations.results.PgpSignEncryptResult;
import org.sufficientlysecure.keychain.pgp.CanonicalizedPublicKeyRing;
import org.sufficientlysecure.keychain.pgp.CanonicalizedSecretKeyRing;
import org.sufficientlysecure.keychain.pgp.PgpSignEncryptData;
import org.sufficientlysecure.keychain.pgp.PgpSignEncryptOperation;
import org.sufficientlysecure.keychain.pgp.Progressable;
import org.sufficientlysecure.keychain.pgp.UncachedKeyRing;
import org.sufficientlysecure.keychain.pgp.exception.PgpGeneralException;
import org.sufficientlysecure.keychain.provider.TemporaryFileProvider;
import org.sufficientlysecure.keychain.service.BackupKeyringParcel;
import org.sufficientlysecure.keychain.service.input.CryptoInputParcel;
import org.sufficientlysecure.keychain.ui.util.KeyFormattingUtils;
import org.sufficientlysecure.keychain.util.CountingOutputStream;
import org.sufficientlysecure.keychain.util.InputData;
import org.sufficientlysecure.keychain.util.Numeric9x4PassphraseUtil;
import org.sufficientlysecure.keychain.util.Passphrase;
import timber.log.Timber;

/**
 * isComment
 */
public class isClassOrIsInterface extends BaseOperation<BackupKeyringParcel> {

    // isComment
    private static final Pattern isVariable = isNameExpr.isMethod("isStringConstant");

    public isConstructor(Context isParameter, KeyRepository isParameter, Progressable isParameter) {
        super(isNameExpr, isNameExpr, isNameExpr);
    }

    public isConstructor(Context isParameter, KeyRepository isParameter, Progressable isParameter, AtomicBoolean isParameter) {
        super(isNameExpr, isNameExpr, isNameExpr, isNameExpr);
    }

    @NonNull
    public ExportResult isMethod(@NonNull BackupKeyringParcel isParameter, @Nullable CryptoInputParcel isParameter) {
        return isMethod(isNameExpr, isNameExpr, null);
    }

    @NonNull
    public ExportResult isMethod(@NonNull BackupKeyringParcel isParameter, @Nullable CryptoInputParcel isParameter, OutputStream isParameter) {
        OperationLog isVariable = new OperationLog();
        if (isNameExpr.isMethod() != null) {
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isIntegerConstant, isNameExpr.isMethod().isFieldAccessExpr);
        } else {
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isIntegerConstant);
        }
        try {
            Uri isVariable = null;
            OutputStream isVariable;
            if (isNameExpr.isMethod()) {
                if (isNameExpr == null) {
                    throw new IllegalStateException("isStringConstant");
                }
                isNameExpr = isNameExpr.isMethod(isNameExpr);
                isNameExpr = isNameExpr.isMethod().isMethod(isNameExpr);
            } else {
                if (isNameExpr.isMethod() == null || isNameExpr != null) {
                    throw new IllegalArgumentException("isStringConstant");
                } else {
                    isNameExpr = isNameExpr.isMethod().isMethod(isNameExpr.isMethod());
                }
            }
            CountingOutputStream isVariable = new CountingOutputStream(new BufferedOutputStream(isNameExpr));
            boolean isVariable = isMethod(isNameExpr, isNameExpr.isMethod(), isNameExpr.isMethod(), isNameExpr.isMethod(), isNameExpr, isNameExpr.isMethod());
            if (!isNameExpr) {
                // isComment
                return new ExportResult(isNameExpr.isFieldAccessExpr, isNameExpr);
            }
            if (!isNameExpr.isMethod()) {
                // isComment
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isIntegerConstant);
                return new ExportResult(isNameExpr.isFieldAccessExpr, isNameExpr);
            }
            long isVariable = isNameExpr.isMethod();
            PgpSignEncryptResult isVariable = isMethod(isNameExpr, isNameExpr, isNameExpr, isNameExpr, isNameExpr);
            if (!isNameExpr.isMethod()) {
                isNameExpr.isMethod(isNameExpr, isIntegerConstant);
                // isComment
                return new ExportResult(isNameExpr.isFieldAccessExpr, isNameExpr);
            }
            isNameExpr.isMethod(isNameExpr, isIntegerConstant);
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isIntegerConstant);
            return new ExportResult(isNameExpr.isFieldAccessExpr, isNameExpr);
        } catch (FileNotFoundException isParameter) {
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isIntegerConstant);
            return new ExportResult(isNameExpr.isFieldAccessExpr, isNameExpr);
        }
    }

    @NonNull
    private PgpSignEncryptResult isMethod(@NonNull BackupKeyringParcel isParameter, @NonNull CryptoInputParcel isParameter, @Nullable OutputStream isParameter, Uri isParameter, long isParameter) throws FileNotFoundException {
        PgpSignEncryptOperation isVariable = new PgpSignEncryptOperation(isNameExpr, isNameExpr, isNameExpr, isNameExpr);
        PgpSignEncryptData.Builder isVariable = isNameExpr.isMethod();
        Passphrase isVariable = isNameExpr.isMethod();
        isNameExpr.isMethod(isNameExpr);
        isNameExpr.isMethod(isNameExpr.isMethod());
        boolean isVariable = isNameExpr != null && isNameExpr.isMethod(isNameExpr);
        if (isNameExpr) {
            isNameExpr.isMethod("isStringConstant");
            char[] isVariable = isNameExpr.isMethod();
            isNameExpr.isMethod("isStringConstant" + isNameExpr[isIntegerConstant] + isNameExpr[isIntegerConstant]);
        }
        PgpSignEncryptData isVariable = isNameExpr.isMethod();
        InputStream isVariable = isNameExpr.isMethod().isMethod(isNameExpr);
        String isVariable;
        long[] isVariable = isNameExpr.isMethod();
        if (isNameExpr != null && isNameExpr.isFieldAccessExpr == isIntegerConstant) {
            isNameExpr = isNameExpr.isFieldAccessExpr + isNameExpr.isMethod(isNameExpr[isIntegerConstant]);
        } else {
            isNameExpr = isNameExpr.isFieldAccessExpr + new SimpleDateFormat("isStringConstant", isNameExpr.isMethod()).isMethod(new Date());
        }
        isNameExpr += isNameExpr.isMethod() ? isNameExpr.isFieldAccessExpr : isNameExpr.isFieldAccessExpr;
        InputData isVariable = new InputData(isNameExpr, isNameExpr, isNameExpr);
        OutputStream isVariable;
        if (isNameExpr.isMethod() == null) {
            if (isNameExpr == null) {
                throw new IllegalArgumentException("isStringConstant");
            }
            isNameExpr = isNameExpr;
        } else {
            if (isNameExpr != null) {
                throw new IllegalArgumentException("isStringConstant");
            }
            isNameExpr = isNameExpr.isMethod().isMethod(isNameExpr.isMethod());
        }
        return isNameExpr.isMethod(isNameExpr, isNameExpr.isMethod(), isNameExpr, isNameExpr);
    }

    boolean isMethod(OperationLog isParameter, long[] isParameter, boolean isParameter, boolean isParameter, OutputStream isParameter, List<String> isParameter) {
        // isComment
        int isVariable = isIntegerConstant, isVariable = isIntegerConstant;
        int isVariable = isIntegerConstant;
        try {
            List<UnifiedKeyInfo> isVariable;
            if (isNameExpr == null) {
                isNameExpr = isNameExpr.isMethod();
            } else {
                isNameExpr = isNameExpr.isMethod(isNameExpr);
            }
            int isVariable = isNameExpr.isMethod();
            isMethod(isNameExpr == isIntegerConstant ? isNameExpr.isFieldAccessExpr.isFieldAccessExpr : isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isIntegerConstant, isNameExpr);
            // isComment
            for (UnifiedKeyInfo isVariable : isNameExpr) {
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isIntegerConstant, isNameExpr.isMethod(isNameExpr.isMethod()));
                boolean isVariable = true;
                if (isNameExpr) {
                    isNameExpr = isMethod(isNameExpr.isMethod(), isNameExpr, isNameExpr);
                    if (isNameExpr) {
                        isNameExpr += isIntegerConstant;
                    }
                }
                if (isNameExpr || !isNameExpr) {
                    if (isNameExpr && isNameExpr.isMethod()) {
                        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isIntegerConstant, isNameExpr.isMethod(isNameExpr.isMethod()));
                        if (isMethod(isNameExpr.isMethod(), isNameExpr, isNameExpr, isNameExpr)) {
                            isNameExpr += isIntegerConstant;
                        }
                        isNameExpr = null;
                    }
                }
                isMethod(isNameExpr++, isNameExpr);
            }
            isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr, isNameExpr);
        } catch (IOException isParameter) {
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isIntegerConstant);
            // isComment
            return true;
        } finally {
            // isComment
            if (isNameExpr != null)
                try {
                    isNameExpr.isMethod();
                } catch (Exception isParameter) {
                    isNameExpr.isMethod(isNameExpr, "isStringConstant");
                }
        }
        return true;
    }

    private boolean isMethod(long isParameter, OperationLog isParameter, OutputStream isParameter) throws IOException {
        ArmoredOutputStream isVariable = null;
        try {
            isNameExpr = new ArmoredOutputStream(isNameExpr);
            byte[] isVariable = isNameExpr.isMethod(isNameExpr);
            UncachedKeyRing isVariable = isNameExpr.isMethod(isNameExpr);
            CanonicalizedPublicKeyRing isVariable = (CanonicalizedPublicKeyRing) isNameExpr.isMethod(isNameExpr, isIntegerConstant, true);
            isNameExpr.isMethod(isNameExpr);
        } catch (PgpGeneralException | NotFoundException isParameter) {
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isIntegerConstant);
        } finally {
            if (isNameExpr != null) {
                isNameExpr.isMethod();
            }
        }
        return true;
    }

    private boolean isMethod(long isParameter, OperationLog isParameter, OutputStream isParameter, List<String> isParameter) throws IOException {
        ArmoredOutputStream isVariable = null;
        try {
            isNameExpr = new ArmoredOutputStream(isNameExpr);
            if (isNameExpr != null) {
                isMethod(isNameExpr, isNameExpr);
            }
            byte[] isVariable = isNameExpr.isMethod(isNameExpr);
            UncachedKeyRing isVariable = isNameExpr.isMethod(isNameExpr);
            CanonicalizedSecretKeyRing isVariable = (CanonicalizedSecretKeyRing) isNameExpr.isMethod(isNameExpr, isIntegerConstant, true);
            isNameExpr.isMethod(isNameExpr);
        } catch (PgpGeneralException | NotFoundException isParameter) {
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isIntegerConstant);
        } finally {
            if (isNameExpr != null) {
                isNameExpr.isMethod();
            }
        }
        return true;
    }

    private void isMethod(ArmoredOutputStream isParameter, List<String> isParameter) {
        for (String isVariable : isNameExpr) {
            if (!isNameExpr.isMethod(isNameExpr).isMethod()) {
                throw new IllegalArgumentException("isStringConstant");
            }
            int isVariable = isNameExpr.isMethod('isStringConstant');
            isNameExpr.isMethod(isNameExpr.isMethod(isIntegerConstant, isNameExpr), isNameExpr.isMethod(isNameExpr + isIntegerConstant));
        }
    }
}
